services/email_service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.2
pydantic==2.11.7
pydantic_core==2.33.2

aiokafka==0.12.0
aiohttp==3.12.15
confluent-kafka==2.12.2
kafka-python==2.2.15

clickhouse-driver==0.2.6
psycopg2-binary==2.9.10
redis==6.4.0

prometheus_client==0.23.1
loguru==0.7.3
dynaconf==3.2.11
python-dotenv==1.1.1

aiofiles==24.1.0
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.8.3
charset-normalizer==3.4.4
click==8.3.0
h11==0.16.0
idna==3.10
lz4==4.4.4
multidict==6.6.3
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
shortuuid==1.0.13
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
zstandard==0.24.0
clickhouse_connect
python-multipart